Add `-tags` parameter (preview)
This is motivated by the fact that `-search` may download a version of a video that could be not the one wanted e.g. `uncensored`.

This unfortunately add some back and forth to fetch the page detail, but this also could be used as a way to globally search and select some video specifically. It can technically be done with `-urls` or `-playlist` but this as a finer control.